Lau - Correcciones UI

+++ Docs/07-Implementacion/Source/trunk/EDUAR_actual/EDUAR/EDUAR_DataTransferObject/Entities/Package Administracion/Pagina.cs
@@ -9,7 +9,7 @@ namespace EDUAR_Entities
     [Serializable]
     public class Pagina : DTBase
     {
-        public int idPagina { get; set; }
+        public long idPagina { get; set; }
         public string titulo { get; set; }
         public string url { get; set; }
     }
